assets/logs/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
jupyter>=1.0.0
kagglehub>=0.2.0